refactor: extract date/height picker into StartHeightPicker widget
