+++ lib/pages/add_wallet_views/restore_wallet_view/restore_options_view/restore_options_view.dart
@@ -8,70 +8,38 @@
  *
  */
 
-import 'dart:async';
-import 'dart:convert';
-import 'dart:io';
-
 import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:logger/logger.dart';
 import 'package:tuple/tuple.dart';
-import 'package:wakelock_plus/wakelock_plus.dart';
 
-import '../../../../models/keys/view_only_wallet_data.dart';
-import '../../../../pages_desktop_specific/desktop_home_view.dart';
 import '../../../../pages_desktop_specific/my_stack_view/exit_to_my_stack_button.dart';
-import '../../../../providers/global/secure_store_provider.dart';
-import '../../../../providers/providers.dart';
 import '../../../../providers/ui/verify_recovery_phrase/mnemonic_word_count_state_provider.dart';
 import '../../../../themes/stack_colors.dart';
-import '../../../../utilities/address_utils.dart';
 import '../../../../utilities/assets.dart';
 import '../../../../utilities/constants.dart';
-import '../../../../utilities/format.dart';
-import '../../../../utilities/logger.dart';
 import '../../../../utilities/text_styles.dart';
 import '../../../../utilities/util.dart';
 import '../../../../wallets/crypto_currency/crypto_currency.dart';
 import '../../../../wallets/crypto_currency/interfaces/view_only_option_currency_interface.dart';
 import '../../../../wallets/crypto_currency/intermediate/cryptonote_currency.dart';
-import '../../../../wallets/isar/models/wallet_info.dart';
-import '../../../../wallets/wallet/intermediate/cryptonote_wallet.dart';
-import '../../../../wallets/wallet/wallet.dart';
 import '../../../../widgets/conditional_parent.dart';
 import '../../../../widgets/custom_buttons/app_bar_icon_button.dart';
-import '../../../../widgets/custom_buttons/blue_text_button.dart';
-import '../../../../widgets/date_picker/date_picker.dart';
 import '../../../../widgets/desktop/desktop_app_bar.dart';
 import '../../../../widgets/desktop/desktop_scaffold.dart';
 import '../../../../widgets/expandable.dart';
-import '../../../../widgets/icon_widgets/x_icon.dart';
-import '../../../../widgets/options.dart';
 import '../../../../widgets/rounded_white_container.dart';
 import '../../../../widgets/stack_text_field.dart';
-import '../../../../widgets/textfield_icon_button.dart';
+import '../../../../widgets/start_height_picker.dart';
 import '../../../../widgets/toggle.dart';
-import '../../../../wl_gen/interfaces/cs_monero_interface.dart';
-import '../../../../wl_gen/interfaces/cs_salvium_interface.dart';
-import '../../../../wl_gen/interfaces/cs_wownero_interface.dart';
-import '../../../home_view/home_view.dart';
 import '../../create_or_restore_wallet_view/sub_widgets/coin_image.dart';
-import '../confirm_recovery_dialog.dart';
 import '../restore_view_only_wallet_view.dart';
 import '../restore_wallet_view.dart';
 import '../sub_widgets/mnemonic_word_count_select_sheet.dart';
-import '../sub_widgets/restore_failed_dialog.dart';
-import '../sub_widgets/restore_succeeded_dialog.dart';
-import '../sub_widgets/restoring_dialog.dart';
 import 'sub_widgets/mobile_mnemonic_length_selector.dart';
-import 'sub_widgets/restore_from_date_picker.dart';
 import 'sub_widgets/restore_options_next_button.dart';
 import 'sub_widgets/restore_options_platform_layout.dart';
-
-final _pIsUsingDate = StateProvider.autoDispose((_) => true);
 
 class RestoreOptionsView extends ConsumerStatefulWidget {
   const RestoreOptionsView({
@@ -94,17 +62,12 @@ class _RestoreOptionsViewState extends ConsumerState<RestoreOptionsView> {
   late final CryptoCurrency coin;
   late final bool isDesktop;
 
-  late TextEditingController _dateController;
-  late TextEditingController _blockHeightController;
-  late FocusNode _blockHeightFocusNode;
   late FocusNode textFieldFocusNode;
   late final FocusNode passwordFocusNode;
   late final TextEditingController passwordController;
+  late final StartHeightPickerController _heightController;
 
-  bool _hasBlockHeight = false;
-  DateTime? _restoreFromDate;
   bool hidePassword = true;
-  WalletUriData? _uriData;
 
   @override
   void initState() {
@@ -113,33 +76,18 @@ class _RestoreOptionsViewState extends ConsumerState<RestoreOptionsView> {
     coin = widget.coin;
     isDesktop = Util.isDesktop;
 
-    _dateController = TextEditingController();
     textFieldFocusNode = FocusNode();
     passwordController = TextEditingController();
     passwordFocusNode = FocusNode();
-    _blockHeightController = TextEditingController();
-    _blockHeightFocusNode = FocusNode();
-
-    _blockHeightController.addListener(() {
-      WidgetsBinding.instance.addPostFrameCallback((_) {
-        if (mounted) {
-          if (!ref.read(_pIsUsingDate)) {
-            setState(() {
-              _hasBlockHeight = _blockHeightController.text.isNotEmpty;
-            });
-          }
-        }
-      });
-    });
+    _heightController = StartHeightPickerController();
   }
 
   @override
   void dispose() {
-    _dateController.dispose();
-    _blockHeightController.dispose();
     textFieldFocusNode.dispose();
     passwordController.dispose();
     passwordFocusNode.dispose();
+    _heightController.dispose();
     super.dispose();
   }
 
@@ -157,66 +105,31 @@ class _RestoreOptionsViewState extends ConsumerState<RestoreOptionsView> {
       }
 
       if (mounted) {
-        int height = 0;
-        if (ref.read(_pIsUsingDate)) {
-          height = getBlockHeightFromDate(_restoreFromDate);
+        final int height = _heightController.height;
+        if (!_showViewOnlyOption) {
+          await Navigator.of(context).pushNamed(
+            RestoreWalletView.routeName,
+            arguments: Tuple5(
+              walletName,
+              coin,
+              ref.read(mnemonicWordCountStateProvider.state).state,
+              height,
+              passwordController.text,
+            ),
+          );
         } else {
-          height = int.tryParse(_blockHeightController.text) ?? 0;
-        }
-        switch (_restoreMode) {
-          case 0: // Seed
-            await Navigator.of(context).pushNamed(
-              RestoreWalletView.routeName,
-              arguments: Tuple5(
-                walletName,
-                coin,
-                ref.read(mnemonicWordCountStateProvider.state).state,
-                height,
-                passwordController.text,
-              ),
-            );
-            break;
-          case 1: // View Only
-            await Navigator.of(context).pushNamed(
-              RestoreViewOnlyWalletView.routeName,
-              arguments: (
-                walletName: walletName,
-                coin: coin,
-                restoreBlockHeight: height,
-              ),
-            );
-            break;
-          case 2: // URI
-            await _attemptUriRestore(height);
-            break;
+          await Navigator.of(context).pushNamed(
+            RestoreViewOnlyWalletView.routeName,
+            arguments: (
+              walletName: walletName,
+              coin: coin,
+              restoreBlockHeight: height,
+            ),
+          );
         }
       }
     } finally {
       _nextLock = false;
-    }
-  }
-
-  Future<void> chooseDate() async {
-    // check and hide keyboard
-    if (FocusScope.of(context).hasFocus) {
-      FocusScope.of(context).unfocus();
-      await Future<void>.delayed(const Duration(milliseconds: 125));
-    }
-
-    if (mounted) {
-      final date = (await showSWDatePicker(context))?.first;
-      if (date != null) {
-        _restoreFromDate = date;
-        _dateController.text = Format.formatDate(date);
-      }
-    }
-  }
-
-  Future<void> chooseDesktopDate() async {
-    final date = (await showSWDatePicker(context))?.first;
-    if (date != null) {
-      _restoreFromDate = date;
-      _dateController.text = Format.formatDate(date);
     }
   }
 
@@ -235,238 +148,7 @@ class _RestoreOptionsViewState extends ConsumerState<RestoreOptionsView> {
     );
   }
 
-  int getBlockHeightFromDate(DateTime? date) {
-    try {
-      int height = 0;
-      if (date != null) {
-        if (widget.coin is Monero) {
-          height = csMonero.getHeightByDate(date);
-        }
-        if (widget.coin is Wownero) {
-          height = csWownero.getHeightByDate(date);
-        }
-        if (widget.coin is Salvium) {
-          height = csSalvium.getHeightByDate(
-            DateTime.now().subtract(const Duration(days: 7)),
-          );
-        }
-        if (height < 0) {
-          height = 0;
-        }
-
-        if (widget.coin is Epiccash) {
-          final int secondsSinceEpoch = date.millisecondsSinceEpoch ~/ 1000;
-          const int epicCashFirstBlock = 1565370278;
-          const double overestimateSecondsPerBlock = 61;
-          final int chosenSeconds = secondsSinceEpoch - epicCashFirstBlock;
-          final int approximateHeight =
-              chosenSeconds ~/ overestimateSecondsPerBlock;
-
-          height = approximateHeight;
-          if (height < 0) {
-            height = 0;
-          }
-        }
-      } else {
-        height = 0;
-      }
-      return height;
-    } catch (e) {
-      Logging.instance.log(
-        Level.info,
-        "Error getting block height from date: $e",
-      );
-      return 0;
-    }
-  }
-
-  Future<void> _attemptUriRestore(int fallbackHeight) async {
-    final data = _uriData;
-    if (data == null) return;
-
-    if (!isDesktop) {
-      FocusScope.of(context).unfocus();
-      await Future<void>.delayed(const Duration(milliseconds: 100));
-    }
-
-    if (!mounted) return;
-
-    await showDialog<dynamic>(
-      context: context,
-      useSafeArea: false,
-      barrierDismissible: true,
-      builder: (context) {
-        return ConfirmRecoveryDialog(
-          onConfirm: () => _doUriRestore(data, fallbackHeight),
-        );
-      },
-    );
-  }
-
-  Future<void> _doUriRestore(WalletUriData data, int fallbackHeight) async {
-    if (!Platform.isLinux && !isDesktop) await WakelockPlus.enable();
-
-    final restoreHeight = data.height ?? fallbackHeight;
-
-    try {
-      final Map<String, dynamic> otherDataJson;
-      if (data.seed != null) {
-        otherDataJson = {};
-      } else if (data.isViewOnly) {
-        otherDataJson = {
-          WalletInfoKeys.isViewOnlyKey: true,
-          WalletInfoKeys.viewOnlyTypeIndexKey:
-              ViewOnlyWalletType.cryptonote.index,
-        };
-      } else {
-        otherDataJson = {WalletInfoKeys.isRestoredFromKeysKey: true};
-      }
-
-      final info = WalletInfo.createNew(
-        coin: coin,
-        name: walletName,
-        restoreHeight: restoreHeight,
-        otherDataJsonString: jsonEncode(otherDataJson),
-      );
-
-      bool isRestoring = true;
-      if (mounted) {
-        unawaited(
-          showDialog<dynamic>(
-            context: context,
-            useSafeArea: false,
-            barrierDismissible: false,
-            builder: (context) {
-              return RestoringDialog(
-                onCancel: () async {
-                  isRestoring = false;
-                  await ref
-                      .read(pWallets)
-                      .deleteWallet(info, ref.read(secureStoreProvider));
-                },
-              );
-            },
-          ),
-        );
-      }
-
-      try {
-        var node = ref
-            .read(nodeServiceChangeNotifierProvider)
-            .getPrimaryNodeFor(currency: coin);
-
-        if (node == null) {
-          node = coin.defaultNode(isPrimary: true);
-          await ref
-              .read(nodeServiceChangeNotifierProvider)
-              .save(node, null, false);
-        }
-
-        final Wallet wallet;
-        if (data.seed != null) {
-          wallet = await Wallet.create(
-            walletInfo: info,
-            mainDB: ref.read(mainDBProvider),
-            secureStorageInterface: ref.read(secureStoreProvider),
-            nodeService: ref.read(nodeServiceChangeNotifierProvider),
-            prefs: ref.read(prefsChangeNotifierProvider),
-            mnemonic: data.seed,
-          );
-        } else if (data.isViewOnly) {
-          final viewOnlyData = CryptonoteViewOnlyWalletData(
-            walletId: info.walletId,
-            address: data.address ?? "",
-            privateViewKey: data.viewKey!,
-          );
-          wallet = await Wallet.create(
-            walletInfo: info,
-            mainDB: ref.read(mainDBProvider),
-            secureStorageInterface: ref.read(secureStoreProvider),
-            nodeService: ref.read(nodeServiceChangeNotifierProvider),
-            prefs: ref.read(prefsChangeNotifierProvider),
-            viewOnlyData: viewOnlyData,
-          );
-        } else {
-          final keysRestoreData = jsonEncode({
-            "address": data.address ?? "",
-            "viewKey": data.viewKey!,
-            "spendKey": data.spendKey!,
-          });
-          wallet = await Wallet.create(
-            walletInfo: info,
-            mainDB: ref.read(mainDBProvider),
-            secureStorageInterface: ref.read(secureStoreProvider),
-            nodeService: ref.read(nodeServiceChangeNotifierProvider),
-            prefs: ref.read(prefsChangeNotifierProvider),
-            keysRestoreData: keysRestoreData,
-          );
-        }
-
-        if (wallet is CryptonoteWallet) {
-          await wallet.init(isRestore: true);
-        } else {
-          await wallet.init();
-        }
-
-        await wallet.recover(isRescan: false);
-
-        if (mounted) {
-          await wallet.info.setMnemonicVerified(
-            isar: ref.read(mainDBProvider).isar,
-          );
-
-          if (ref.read(pDuress)) {
-            await wallet.info.updateDuressVisibilityStatus(
-              isDuressVisible: true,
-              isar: ref.read(mainDBProvider).isar,
-            );
-          }
-
-          ref.read(pWallets).addWallet(wallet);
-
-          if (mounted) {
-            if (isDesktop) {
-              Navigator.of(
-                context,
-              ).popUntil(ModalRoute.withName(DesktopHomeView.routeName));
-            } else {
-              unawaited(
-                Navigator.of(
-                  context,
-                ).pushNamedAndRemoveUntil(HomeView.routeName, (route) => false),
-              );
-            }
-
-            await showDialog<dynamic>(
-              context: context,
-              useSafeArea: false,
-              barrierDismissible: true,
-              builder: (context) => const RestoreSucceededDialog(),
-            );
-          }
-        }
-      } catch (e) {
-        if (mounted && isRestoring) {
-          Navigator.pop(context);
-          await showDialog<dynamic>(
-            context: context,
-            useSafeArea: false,
-            barrierDismissible: true,
-            builder: (context) => RestoreFailedDialog(
-              errorMessage: e.toString(),
-              walletId: info.walletId,
-              walletName: info.name,
-            ),
-          );
-        }
-      }
-    } finally {
-      if (!Platform.isLinux && !isDesktop) await WakelockPlus.disable();
-    }
-  }
-
-  // 0 = Seed, 1 = View Only, 2 = URI (Monero only)
-  int _restoreMode = 0;
+  bool _showViewOnlyOption = false;
 
   @override
   Widget build(BuildContext context) {
@@ -518,98 +200,52 @@ class _RestoreOptionsViewState extends ConsumerState<RestoreOptionsView> {
                 SizedBox(
                   height: isDesktop ? 56 : 48,
                   width: isDesktop ? 490 : null,
-                  child: coin is Monero
-                      ? Options(
-                          key: UniqueKey(),
-                          texts: const ["Seed", "View Only", "URI"],
-                          onColor: Theme.of(
-                            context,
-                          ).extension<StackColors>()!.popupBG,
-                          offColor: Theme.of(
-                            context,
-                          ).extension<StackColors>()!.textFieldDefaultBG,
-                          selectedIndex: _restoreMode,
-                          onValueChanged: (value) {
-                            setState(() {
-                              _restoreMode = value;
-                            });
-                          },
-                          decoration: BoxDecoration(
-                            color: Colors.transparent,
-                            borderRadius: BorderRadius.circular(
-                              Constants.size.circularBorderRadius,
-                            ),
-                          ),
-                        )
-                      : Toggle(
-                          key: UniqueKey(),
-                          onText: "Seed",
-                          offText: "View Only",
-                          onColor: Theme.of(
-                            context,
-                          ).extension<StackColors>()!.popupBG,
-                          offColor: Theme.of(
-                            context,
-                          ).extension<StackColors>()!.textFieldDefaultBG,
-                          isOn: _restoreMode == 1,
-                          onValueChanged: (value) {
-                            setState(() {
-                              _restoreMode = value ? 1 : 0;
-                            });
-                          },
-                          decoration: BoxDecoration(
-                            color: Colors.transparent,
-                            borderRadius: BorderRadius.circular(
-                              Constants.size.circularBorderRadius,
-                            ),
-                          ),
-                        ),
+                  child: Toggle(
+                    key: UniqueKey(),
+                    onText: "Seed",
+                    offText: "View Only",
+                    onColor: Theme.of(
+                      context,
+                    ).extension<StackColors>()!.popupBG,
+                    offColor: Theme.of(
+                      context,
+                    ).extension<StackColors>()!.textFieldDefaultBG,
+                    isOn: _showViewOnlyOption,
+                    onValueChanged: (value) {
+                      setState(() {
+                        _showViewOnlyOption = value;
+                      });
+                    },
+                    decoration: BoxDecoration(
+                      color: Colors.transparent,
+                      borderRadius: BorderRadius.circular(
+                        Constants.size.circularBorderRadius,
+                      ),
+                    ),
+                  ),
                 ),
               if (coin is ViewOnlyOptionCurrencyInterface)
                 SizedBox(height: isDesktop ? 40 : 24),
-              if (_restoreMode == 1)
-                ViewOnlyRestoreOption(
-                  coin: coin,
-                  dateController: _dateController,
-                  dateChooserFunction: isDesktop
-                      ? chooseDesktopDate
-                      : chooseDate,
-                  blockHeightController: _blockHeightController,
-                  blockHeightFocusNode: _blockHeightFocusNode,
-                )
-              else if (_restoreMode == 2)
-                UriRestoreOption(
-                  coin: coin,
-                  dateController: _dateController,
-                  dateChooserFunction: isDesktop
-                      ? chooseDesktopDate
-                      : chooseDate,
-                  blockHeightController: _blockHeightController,
-                  blockHeightFocusNode: _blockHeightFocusNode,
-                  onParsed: (data) => setState(() => _uriData = data),
-                )
-              else
-                SeedRestoreOption(
-                  coin: coin,
-                  dateController: _dateController,
-                  blockHeightController: _blockHeightController,
-                  blockHeightFocusNode: _blockHeightFocusNode,
-                  pwController: passwordController,
-                  pwFocusNode: passwordFocusNode,
-                  dateChooserFunction: isDesktop
-                      ? chooseDesktopDate
-                      : chooseDate,
-                  chooseMnemonicLength: chooseMnemonicLength,
-                ),
+              _showViewOnlyOption
+                  ? ViewOnlyRestoreOption(
+                      coin: coin,
+                      heightController: _heightController,
+                    )
+                  : SeedRestoreOption(
+                      coin: coin,
+                      heightController: _heightController,
+                      pwController: passwordController,
+                      pwFocusNode: passwordFocusNode,
+                      chooseMnemonicLength: chooseMnemonicLength,
+                    ),
               if (!isDesktop) const Spacer(flex: 3),
               SizedBox(height: isDesktop ? 32 : 12),
-              RestoreOptionsNextButton(
-                isDesktop: isDesktop,
-                onPressed: _restoreMode == 2
-                    ? (_uriData != null ? nextPressed : null)
-                    : ref.watch(_pIsUsingDate) || _hasBlockHeight
-                    ? nextPressed
-                    : null,
+              ListenableBuilder(
+                listenable: _heightController,
+                builder: (context, _) => RestoreOptionsNextButton(
+                  isDesktop: isDesktop,
+                  onPressed: _heightController.canProceed ? nextPressed : null,
+                ),
               ),
               if (isDesktop) const Spacer(flex: 15),
             ],
@@ -624,23 +260,17 @@ class SeedRestoreOption extends ConsumerStatefulWidget {
   const SeedRestoreOption({
     super.key,
     required this.coin,
-    required this.dateController,
-    required this.blockHeightController,
-    required this.blockHeightFocusNode,
+    required this.heightController,
     required this.pwController,
     required this.pwFocusNode,
-    required this.dateChooserFunction,
     required this.chooseMnemonicLength,
   });
 
   final CryptoCurrency coin;
-  final TextEditingController dateController;
-  final TextEditingController blockHeightController;
-  final FocusNode blockHeightFocusNode;
+  final StartHeightPickerController heightController;
   final TextEditingController pwController;
   final FocusNode pwFocusNode;
 
-  final Future<void> Function() dateChooserFunction;
   final Future<void> Function() chooseMnemonicLength;
 
   @override
@@ -650,7 +280,6 @@ class SeedRestoreOption extends ConsumerStatefulWidget {
 class _SeedRestoreOptionState extends ConsumerState<SeedRestoreOption> {
   bool _hidePassword = true;
   bool _expandedAdvanced = false;
-  bool _blockFieldEmpty = true;
 
   @override
   Widget build(BuildContext context) {
@@ -675,120 +304,13 @@ class _SeedRestoreOptionState extends ConsumerState<SeedRestoreOption> {
       children: [
         if (isCnAnd25 ||
             widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                ref.watch(_pIsUsingDate) ? "Choose start date" : "Block height",
-                style: Util.isDesktop
-                    ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                        color: Theme.of(
-                          context,
-                        ).extension<StackColors>()!.textDark3,
-                      )
-                    : STextStyles.smallMed12(context),
-                textAlign: TextAlign.left,
-              ),
-              CustomTextButton(
-                text: ref.watch(_pIsUsingDate)
-                    ? "Use block height"
-                    : "Use date",
-                onTap: () => ref.read(_pIsUsingDate.notifier).state = !ref.read(
-                  _pIsUsingDate,
-                ),
-              ),
-            ],
+            widget.coin is Mimblewimblecoin) ...[
+          StartHeightPicker(
+            coin: widget.coin,
+            controller: widget.heightController,
           ),
-        if (isCnAnd25 ||
-            widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
-          SizedBox(height: Util.isDesktop ? 16 : 8),
-        if (isCnAnd25 ||
-            widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
-          ref.watch(_pIsUsingDate)
-              ? RestoreFromDatePicker(
-                  onTap: widget.dateChooserFunction,
-                  controller: widget.dateController,
-                )
-              : ClipRRect(
-                  borderRadius: BorderRadius.circular(
-                    Constants.size.circularBorderRadius,
-                  ),
-                  child: TextField(
-                    focusNode: widget.blockHeightFocusNode,
-                    controller: widget.blockHeightController,
-                    keyboardType: TextInputType.number,
-                    inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                    textInputAction: TextInputAction.done,
-                    style: Util.isDesktop
-                        ? STextStyles.desktopTextMedium(
-                            context,
-                          ).copyWith(height: 2)
-                        : STextStyles.field(context),
-                    onChanged: (value) {
-                      setState(() {
-                        _blockFieldEmpty = value.isEmpty;
-                      });
-                    },
-                    decoration:
-                        standardInputDecoration(
-                          "Start scanning from...",
-                          widget.blockHeightFocusNode,
-                          context,
-                        ).copyWith(
-                          suffixIcon: UnconstrainedBox(
-                            child: TextFieldIconButton(
-                              child: Semantics(
-                                label:
-                                    "Clear Block Height Field Button. Clears the block height field",
-                                excludeSemantics: true,
-                                child: !_blockFieldEmpty
-                                    ? XIcon(
-                                        width: Util.isDesktop ? 24 : 16,
-                                        height: Util.isDesktop ? 24 : 16,
-                                      )
-                                    : const SizedBox.shrink(),
-                              ),
-                              onTap: () {
-                                widget.blockHeightController.text = "";
-                                setState(() {
-                                  _blockFieldEmpty = true;
-                                });
-                              },
-                            ),
-                          ),
-                        ),
-                  ),
-                ),
-        if (isCnAnd25 ||
-            widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
-          const SizedBox(height: 8),
-        if (isCnAnd25 ||
-            widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
-          RoundedWhiteContainer(
-            child: Center(
-              child: Text(
-                ref.watch(_pIsUsingDate)
-                    ? "Choose the date you made the wallet (approximate is fine)"
-                    : "Enter the initial block height of the wallet",
-                style: Util.isDesktop
-                    ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                        color: Theme.of(
-                          context,
-                        ).extension<StackColors>()!.textSubtitle1,
-                      )
-                    : STextStyles.smallMed12(context).copyWith(fontSize: 10),
-              ),
-            ),
-          ),
-        if (isCnAnd25 ||
-            widget.coin is Epiccash ||
-            widget.coin is Mimblewimblecoin)
           SizedBox(height: Util.isDesktop ? 24 : 16),
+        ],
         Text(
           "Choose recovery phrase length",
           style: Util.isDesktop
@@ -1002,358 +524,27 @@ class _SeedRestoreOptionState extends ConsumerState<SeedRestoreOption> {
       ],
     );
   }
-
-  @override
-  void initState() {
-    super.initState();
-
-    _blockFieldEmpty = widget.blockHeightController.text.isEmpty;
-  }
 }
 
-class ViewOnlyRestoreOption extends ConsumerStatefulWidget {
+class ViewOnlyRestoreOption extends StatelessWidget {
   const ViewOnlyRestoreOption({
     super.key,
     required this.coin,
-    required this.dateController,
-    required this.dateChooserFunction,
-    required this.blockHeightController,
-    required this.blockHeightFocusNode,
+    required this.heightController,
   });
 
   final CryptoCurrency coin;
-  final TextEditingController dateController;
-  final TextEditingController blockHeightController;
-  final FocusNode blockHeightFocusNode;
-
-  final Future<void> Function() dateChooserFunction;
-
-  @override
-  ConsumerState<ViewOnlyRestoreOption> createState() =>
-      _ViewOnlyRestoreOptionState();
-}
-
-class _ViewOnlyRestoreOptionState extends ConsumerState<ViewOnlyRestoreOption> {
-  bool _blockFieldEmpty = true;
+  final StartHeightPickerController heightController;
 
   @override
   Widget build(BuildContext context) {
-    final showDateOption = widget.coin is CryptonoteCurrency;
+    final showDateOption = coin is CryptonoteCurrency;
     return Column(
       children: [
-        if (showDateOption)
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: [
-              Text(
-                ref.watch(_pIsUsingDate) ? "Choose start date" : "Block height",
-                style: Util.isDesktop
-                    ? STextStyles.desktopTextExtraExtraSmall(context).copyWith(
-                        color: Theme.of(
-                          context,
-                        ).extension<StackColors>()!.textDark3,
-                      )
-                    : STextStyles.smallMed12(context),
-                textAlign: TextAlign.left,
-              ),
-              CustomTextButton(
-                text: ref.watch(_pIsUsingDate)
-                    ? "Use block height"
-                    : "Use date",
-                onTap: () {
-                  ref.read(_pIsUsingDate.notifier).state = !ref.read(
-                    _pIsUsingDate,
-                  );
-                },
-              ),
-            ],
-          ),
-        if (showDateOption) SizedBox(height: Util.isDesktop ? 16 : 8),
-        if (showDateOption)
-          ref.watch(_pIsUsingDate)
-              ? RestoreFromDatePicker(
-                  onTap: widget.dateChooserFunction,
-                  controller: widget.dateController,
-                )
-              : ClipRRect(
-                  borderRadius: BorderRadius.circular(
-                    Constants.size.circularBorderRadius,
-                  ),
-                  child: TextField(
-                    focusNode: widget.blockHeightFocusNode,
-                    controller: widget.blockHeightController,
-                    keyboardType: TextInputType.number,
-                    inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                    textInputAction: TextInputAction.done,
-                    style: Util.isDesktop
-                        ? STextStyles.desktopTextMedium(
-                            context,
-                          ).copyWith(height: 2)
-                        : STextStyles.field(context),
-                    onChanged: (value) {
-                      setState(() {
-                        _blockFieldEmpty = value.isEmpty;
-                      });
-                    },
-                    decoration:
-                        standardInputDecoration(
-                          "Start scanning from...",
-                          widget.blockHeightFocusNode,
-                          context,
-                        ).copyWith(
-                          suffixIcon: UnconstrainedBox(
-                            child: TextFieldIconButton(
-                              child: Semantics(
-                                label:
-                                    "Clear Block Height Field Button. Clears the block height field",
-                                excludeSemantics: true,
-                                child: !_blockFieldEmpty
-                                    ? XIcon(
-                                        width: Util.isDesktop ? 24 : 16,
-                                        height: Util.isDesktop ? 24 : 16,
-                                      )
-                                    : const SizedBox.shrink(),
-                              ),
-                              onTap: () {
-                                widget.blockHeightController.text = "";
-                                setState(() {
-                                  _blockFieldEmpty = true;
-                                });
-                              },
-                            ),
-                          ),
-                        ),
-                  ),
-                ),
-        if (showDateOption) const SizedBox(height: 8),
-        if (showDateOption)
-          RoundedWhiteContainer(
-            child: Center(
-              child: Text(
-                ref.watch(_pIsUsingDate)
-                    ? "Choose the date you made the wallet (approximate is fine)"
-                    : "Enter the initial block height of the wallet",
-                style: Util.isDesktop
-                    ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                        color: Theme.of(
-                          context,
-                        ).extension<StackColors>()!.textSubtitle1,
-                      )
-                    : STextStyles.smallMed12(context).copyWith(fontSize: 10),
-              ),
-            ),
-          ),
-        if (showDateOption) SizedBox(height: Util.isDesktop ? 24 : 16),
-      ],
-    );
-  }
-
-  @override
-  void initState() {
-    super.initState();
-
-    _blockFieldEmpty = widget.blockHeightController.text.isEmpty;
-  }
-}
-
-class UriRestoreOption extends ConsumerStatefulWidget {
-  const UriRestoreOption({
-    super.key,
-    required this.coin,
-    required this.dateController,
-    required this.dateChooserFunction,
-    required this.blockHeightController,
-    required this.blockHeightFocusNode,
-    required this.onParsed,
-  });
-
-  final CryptoCurrency coin;
-  final TextEditingController dateController;
-  final TextEditingController blockHeightController;
-  final FocusNode blockHeightFocusNode;
-  final void Function(WalletUriData?) onParsed;
-
-  final Future<void> Function() dateChooserFunction;
-
-  @override
-  ConsumerState<UriRestoreOption> createState() => _UriRestoreOptionState();
-}
-
-class _UriRestoreOptionState extends ConsumerState<UriRestoreOption> {
-  bool _blockFieldEmpty = true;
-  late final TextEditingController _uriController;
-
-  @override
-  void initState() {
-    super.initState();
-    _blockFieldEmpty = widget.blockHeightController.text.isEmpty;
-    _uriController = TextEditingController();
-  }
-
-  @override
-  void dispose() {
-    _uriController.dispose();
-    super.dispose();
-  }
-
-  void _onUriChanged(String value) {
-    WalletUriData? parsed;
-    try {
-      parsed = WalletUriData.fromUriString(value.trim());
-    } catch (_) {
-      parsed = null;
-    }
-
-    // If the URI contains a height, switch to block height mode and populate.
-    if (parsed?.height != null) {
-      ref.read(_pIsUsingDate.notifier).state = false;
-      widget.blockHeightController.text = parsed!.height.toString();
-    }
-
-    widget.onParsed(parsed);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Text(
-          "Paste wallet URI",
-          style: Util.isDesktop
-              ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                  color: Theme.of(context).extension<StackColors>()!.textDark3,
-                )
-              : STextStyles.smallMed12(context),
-          textAlign: TextAlign.left,
-        ),
-        SizedBox(height: Util.isDesktop ? 16 : 8),
-        ClipRRect(
-          borderRadius: BorderRadius.circular(
-            Constants.size.circularBorderRadius,
-          ),
-          child: TextField(
-            controller: _uriController,
-            style: Util.isDesktop
-                ? STextStyles.desktopTextMedium(context).copyWith(height: 2)
-                : STextStyles.field(context),
-            decoration:
-                standardInputDecoration(
-                  "monero_wallet:<address>?seed=...",
-                  FocusNode(),
-                  context,
-                ).copyWith(
-                  suffixIcon: UnconstrainedBox(
-                    child: TextFieldIconButton(
-                      child: _uriController.text.isNotEmpty
-                          ? XIcon(
-                              width: Util.isDesktop ? 24 : 16,
-                              height: Util.isDesktop ? 24 : 16,
-                            )
-                          : const SizedBox.shrink(),
-                      onTap: () {
-                        _uriController.clear();
-                        _onUriChanged("");
-                      },
-                    ),
-                  ),
-                ),
-            maxLines: 3,
-            minLines: 1,
-            onChanged: _onUriChanged,
-          ),
-        ),
-        SizedBox(height: Util.isDesktop ? 24 : 16),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Text(
-              ref.watch(_pIsUsingDate) ? "Choose start date" : "Block height",
-              style: Util.isDesktop
-                  ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                      color: Theme.of(
-                        context,
-                      ).extension<StackColors>()!.textDark3,
-                    )
-                  : STextStyles.smallMed12(context),
-              textAlign: TextAlign.left,
-            ),
-            CustomTextButton(
-              text: ref.watch(_pIsUsingDate) ? "Use block height" : "Use date",
-              onTap: () => ref.read(_pIsUsingDate.notifier).state = !ref.read(
-                _pIsUsingDate,
-              ),
-            ),
-          ],
-        ),
-        SizedBox(height: Util.isDesktop ? 16 : 8),
-        ref.watch(_pIsUsingDate)
-            ? RestoreFromDatePicker(
-                onTap: widget.dateChooserFunction,
-                controller: widget.dateController,
-              )
-            : ClipRRect(
-                borderRadius: BorderRadius.circular(
-                  Constants.size.circularBorderRadius,
-                ),
-                child: TextField(
-                  focusNode: widget.blockHeightFocusNode,
-                  controller: widget.blockHeightController,
-                  keyboardType: TextInputType.number,
-                  inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-                  textInputAction: TextInputAction.done,
-                  style: Util.isDesktop
-                      ? STextStyles.desktopTextMedium(
-                          context,
-                        ).copyWith(height: 2)
-                      : STextStyles.field(context),
-                  onChanged: (value) {
-                    setState(() {
-                      _blockFieldEmpty = value.isEmpty;
-                    });
-                  },
-                  decoration:
-                      standardInputDecoration(
-                        "Start scanning from...",
-                        widget.blockHeightFocusNode,
-                        context,
-                      ).copyWith(
-                        suffixIcon: UnconstrainedBox(
-                          child: TextFieldIconButton(
-                            child: !_blockFieldEmpty
-                                ? XIcon(
-                                    width: Util.isDesktop ? 24 : 16,
-                                    height: Util.isDesktop ? 24 : 16,
-                                  )
-                                : const SizedBox.shrink(),
-                            onTap: () {
-                              widget.blockHeightController.text = "";
-                              setState(() {
-                                _blockFieldEmpty = true;
-                              });
-                            },
-                          ),
-                        ),
-                      ),
-                ),
-              ),
-        const SizedBox(height: 8),
-        RoundedWhiteContainer(
-          child: Center(
-            child: Text(
-              ref.watch(_pIsUsingDate)
-                  ? "Choose the date you made the wallet (approximate is fine)"
-                  : "Enter the initial block height of the wallet",
-              style: Util.isDesktop
-                  ? STextStyles.desktopTextExtraSmall(context).copyWith(
-                      color: Theme.of(
-                        context,
-                      ).extension<StackColors>()!.textSubtitle1,
-                    )
-                  : STextStyles.smallMed12(context).copyWith(fontSize: 10),
-            ),
-          ),
-        ),
+        if (showDateOption) ...[
+          StartHeightPicker(coin: coin, controller: heightController),
+          SizedBox(height: Util.isDesktop ? 24 : 16),
+        ],
       ],
     );
   }
