+++ lib/widgets/start_height_picker.dart
@@ -1,0 +1,296 @@
+/*
+ * This file is part of Stack Wallet.
+ *
+ * Copyright (c) 2023 Cypher Stack
+ * All Rights Reserved.
+ * The code is distributed under GPLv3 license, see LICENSE file for details.
+ * Generated by Cypher Stack on 2023-05-26
+ *
+ */
+
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:logger/logger.dart';
+
+import '../themes/stack_colors.dart';
+import '../utilities/constants.dart';
+import '../utilities/format.dart';
+import '../utilities/logger.dart';
+import '../utilities/text_styles.dart';
+import '../utilities/util.dart';
+import '../wallets/crypto_currency/crypto_currency.dart';
+import '../wl_gen/interfaces/cs_monero_interface.dart';
+import '../wl_gen/interfaces/cs_salvium_interface.dart';
+import '../wl_gen/interfaces/cs_wownero_interface.dart';
+import 'custom_buttons/blue_text_button.dart';
+import 'date_picker/date_picker.dart';
+import 'icon_widgets/x_icon.dart';
+import '../pages/add_wallet_views/restore_wallet_view/restore_options_view/sub_widgets/restore_from_date_picker.dart';
+import 'rounded_white_container.dart';
+import 'stack_text_field.dart';
+import 'textfield_icon_button.dart';
+
+/// Exposes the current height selection of a [StartHeightPicker] to its parent.
+/// Create one per picker, pass it in, and listen to it to react to changes.
+class StartHeightPickerController extends ChangeNotifier {
+  bool _isUsingDate = true;
+  int _height = 0;
+  bool _hasBlockHeight = false;
+
+  /// Computed block height. 0 when in date mode with no date chosen, or when no
+  /// block height has been entered.
+  int get height => _height;
+
+  /// Whether the picker is in date mode.
+  bool get isUsingDate => _isUsingDate;
+
+  /// Whether a block height has been entered while in block height mode.
+  bool get hasBlockHeight => _hasBlockHeight;
+
+  /// Whether the current state is enough to proceed: date mode is active, or a
+  /// block height has been entered.
+  bool get canProceed => _isUsingDate || _hasBlockHeight;
+
+  // Called by StartHeightPicker whenever its internal state changes.
+  void _update({
+    required bool isUsingDate,
+    required int height,
+    required bool hasBlockHeight,
+  }) {
+    _isUsingDate = isUsingDate;
+    _height = height;
+    _hasBlockHeight = hasBlockHeight;
+    notifyListeners();
+  }
+}
+
+/// Lets the user choose either a calendar date or a block height as the
+/// starting point for a wallet scan or restore. State is managed internally;
+/// the parent receives updates through [StartHeightPickerController].
+class StartHeightPicker extends StatefulWidget {
+  const StartHeightPicker({
+    super.key,
+    required this.coin,
+    required this.controller,
+  });
+
+  final CryptoCurrency coin;
+  final StartHeightPickerController controller;
+
+  @override
+  State<StartHeightPicker> createState() => _StartHeightPickerState();
+}
+
+class _StartHeightPickerState extends State<StartHeightPicker> {
+  late final TextEditingController _dateController;
+  late final TextEditingController _blockHeightController;
+  late final FocusNode _blockHeightFocusNode;
+
+  bool _isUsingDate = true;
+  DateTime? _restoreFromDate;
+  bool _blockFieldEmpty = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _dateController = TextEditingController();
+    _blockHeightController = TextEditingController();
+    _blockHeightFocusNode = FocusNode();
+    // Notify after the first frame so a watching ListenableBuilder does not
+    // rebuild during its own build phase.
+    WidgetsBinding.instance.addPostFrameCallback((_) => _notifyController());
+  }
+
+  @override
+  void dispose() {
+    _dateController.dispose();
+    _blockHeightController.dispose();
+    _blockHeightFocusNode.dispose();
+    super.dispose();
+  }
+
+  void _notifyController() {
+    widget.controller._update(
+      isUsingDate: _isUsingDate,
+      height: _currentHeight,
+      hasBlockHeight: !_blockFieldEmpty,
+    );
+  }
+
+  int _getBlockHeightFromDate(DateTime? date) {
+    try {
+      int height = 0;
+      if (date != null) {
+        if (widget.coin is Monero) {
+          height = csMonero.getHeightByDate(date);
+        }
+        if (widget.coin is Wownero) {
+          height = csWownero.getHeightByDate(date);
+        }
+        if (widget.coin is Salvium) {
+          height = csSalvium.getHeightByDate(date);
+        }
+        if (height < 0) {
+          height = 0;
+        }
+
+        if (widget.coin is Epiccash) {
+          final int secondsSinceEpoch = date.millisecondsSinceEpoch ~/ 1000;
+          const int epicCashFirstBlock = 1565370278;
+          const double overestimateSecondsPerBlock = 61;
+          final int chosenSeconds = secondsSinceEpoch - epicCashFirstBlock;
+          final int approximateHeight =
+              chosenSeconds ~/ overestimateSecondsPerBlock;
+
+          height = approximateHeight;
+          if (height < 0) {
+            height = 0;
+          }
+        }
+      } else {
+        height = 0;
+      }
+      return height;
+    } catch (e) {
+      Logging.instance.log(
+        Level.info,
+        "Error getting block height from date: $e",
+      );
+      return 0;
+    }
+  }
+
+  int get _currentHeight {
+    if (_isUsingDate) {
+      return _getBlockHeightFromDate(_restoreFromDate);
+    } else {
+      return int.tryParse(_blockHeightController.text) ?? 0;
+    }
+  }
+
+  Future<void> _chooseDate() async {
+    if (!Util.isDesktop && FocusScope.of(context).hasFocus) {
+      FocusScope.of(context).unfocus();
+      await Future<void>.delayed(const Duration(milliseconds: 125));
+    }
+    if (mounted) {
+      final date = (await showSWDatePicker(context))?.first;
+      if (date != null) {
+        setState(() {
+          _restoreFromDate = date;
+          _dateController.text = Format.formatDate(date);
+        });
+        _notifyController();
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Text(
+              _isUsingDate ? "Choose start date" : "Block height",
+              style: Util.isDesktop
+                  ? STextStyles.desktopTextExtraSmall(context).copyWith(
+                      color: Theme.of(
+                        context,
+                      ).extension<StackColors>()!.textDark3,
+                    )
+                  : STextStyles.smallMed12(context),
+              textAlign: TextAlign.left,
+            ),
+            CustomTextButton(
+              text: _isUsingDate ? "Use block height" : "Use date",
+              onTap: () {
+                setState(() {
+                  _isUsingDate = !_isUsingDate;
+                });
+                _notifyController();
+              },
+            ),
+          ],
+        ),
+        SizedBox(height: Util.isDesktop ? 16 : 8),
+        _isUsingDate
+            ? RestoreFromDatePicker(
+                onTap: _chooseDate,
+                controller: _dateController,
+              )
+            : ClipRRect(
+                borderRadius: BorderRadius.circular(
+                  Constants.size.circularBorderRadius,
+                ),
+                child: TextField(
+                  focusNode: _blockHeightFocusNode,
+                  controller: _blockHeightController,
+                  keyboardType: TextInputType.number,
+                  inputFormatters: [FilteringTextInputFormatter.digitsOnly],
+                  textInputAction: TextInputAction.done,
+                  style: Util.isDesktop
+                      ? STextStyles.desktopTextMedium(
+                          context,
+                        ).copyWith(height: 2)
+                      : STextStyles.field(context),
+                  onChanged: (value) {
+                    setState(() {
+                      _blockFieldEmpty = value.isEmpty;
+                    });
+                    _notifyController();
+                  },
+                  decoration:
+                      standardInputDecoration(
+                        "Start scanning from...",
+                        _blockHeightFocusNode,
+                        context,
+                      ).copyWith(
+                        suffixIcon: UnconstrainedBox(
+                          child: TextFieldIconButton(
+                            child: Semantics(
+                              label:
+                                  "Clear Block Height Field Button. Clears the block height field",
+                              excludeSemantics: true,
+                              child: !_blockFieldEmpty
+                                  ? XIcon(
+                                      width: Util.isDesktop ? 24 : 16,
+                                      height: Util.isDesktop ? 24 : 16,
+                                    )
+                                  : const SizedBox.shrink(),
+                            ),
+                            onTap: () {
+                              _blockHeightController.text = "";
+                              setState(() {
+                                _blockFieldEmpty = true;
+                              });
+                              _notifyController();
+                            },
+                          ),
+                        ),
+                      ),
+                ),
+              ),
+        const SizedBox(height: 8),
+        RoundedWhiteContainer(
+          child: Center(
+            child: Text(
+              _isUsingDate
+                  ? "Choose the date you made the wallet (approximate is fine)"
+                  : "Enter the initial block height of the wallet",
+              style: Util.isDesktop
+                  ? STextStyles.desktopTextExtraSmall(context).copyWith(
+                      color: Theme.of(
+                        context,
+                      ).extension<StackColors>()!.textSubtitle1,
+                    )
+                  : STextStyles.smallMed12(context).copyWith(fontSize: 10),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
